+++ NEWS.md
@@ -1,3 +1,38 @@
+SqlRender 1.19.6
+================
+
+Bugfixes:
+
+1. Now correctly translating `CREATE UNIQUE CLUSTERED INDEX`, `PRIMARY KEY NONCLUSTERED`, and `UPDATE STATISTICS` on Spark.
+
+
+SqlRender 1.19.5
+================
+
+Bugfixes:
+
+1. Now correctly translating `ALTER TABLE ADD` on DuckDB when adding multiple columns at once.
+
+2. Now correctly translating `ALTER TABLE ALTER COLUMN` on DuckDB.
+
+3. Removing parentheses in CTAS with ORDER BY on InterSystems IRIS
+
+
+SqlRender 1.19.4
+================
+
+Bugfixes:
+
+1. Translating `FLOAT` to `DOUBLE` on DataBricks (Spark), because DataBricks `FLOAT` is single precision while SQL Server `FLOAT` is double precision.
+
+SqlRender 1.19.3
+================
+
+Bugfixes:
+
+1. Fix and simplify support for CTEs in DDL statements for InterSystems IRIS.
+
+
 SqlRender 1.19.2
 ================